backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-dotenv>=1.1.0
python-multipart>=0.0.20
openai>=1.68.0
anthropic>=0.49.0
opencv-python-headless>=4.11.0
Pillow>=11.1.0
jieba>=0.42.1
aiofiles>=24.1.0
httpx>=0.28.0
